update platform modal name update
